Split trace state tests
