+++ processor/src/trace/trace_state.rs
@@ -1883,537 +1883,7 @@ impl Deserializable for ExecutionReplay {
 }
 
 #[cfg(feature = "arbitrary")]
-mod arbitrary {
-    use alloc::collections::VecDeque;
-
-    use miden_air::trace::chiplets::hasher::STATE_WIDTH;
-    use proptest::{collection, prelude::*};
-
-    use super::*;
-
-    const MAX_REPLAY_ITEMS: usize = 8;
-
-    fn arb_row_index() -> impl Strategy<Value = RowIndex> {
-        any::<u32>().prop_map(RowIndex::from)
-    }
-
-    // TODO: use any::<MerklePath>() once miden-crypto exposes its impl outside test code:
-    // https://github.com/0xMiden/crypto/issues/1072
-    fn arb_merkle_path() -> impl Strategy<Value = MerklePath> {
-        collection::vec(any::<Word>(), 0..=MAX_REPLAY_ITEMS).prop_map(MerklePath::new)
-    }
-
-    fn arb_vec_deque<T, S>(strategy: S) -> impl Strategy<Value = VecDeque<T>>
-    where
-        T: core::fmt::Debug + 'static,
-        S: Strategy<Value = T> + 'static,
-    {
-        collection::vec(strategy, 0..=MAX_REPLAY_ITEMS).prop_map(VecDeque::from)
-    }
-
-    impl Arbitrary for SystemState {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (arb_row_index(), any::<ContextId>(), any::<Word>(), any::<Word>())
-                .prop_map(|(clk, ctx, fn_hash, pc_transcript_state)| Self {
-                    clk,
-                    ctx,
-                    fn_hash,
-                    pc_transcript_state,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for DecoderState {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (any::<Felt>(), any::<Felt>())
-                .prop_map(|(current_addr, parent_addr)| Self { current_addr, parent_addr })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for StackState {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                any::<[Felt; MIN_STACK_DEPTH]>(),
-                MIN_STACK_DEPTH..=MIN_STACK_DEPTH + 64,
-                any::<Felt>(),
-            )
-                .prop_map(|(stack_top, stack_depth, last_overflow_addr)| Self {
-                    stack_top,
-                    stack_depth,
-                    last_overflow_addr,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for CoreTraceState {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (any::<SystemState>(), any::<DecoderState>(), any::<StackState>())
-                .prop_map(|(system, decoder, stack)| Self { system, decoder, stack })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for CoreTraceFragmentContext {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                any::<CoreTraceState>(),
-                any::<ExecutionReplay>(),
-                any::<ContinuationStack<MastForestId>>(),
-                any::<MastForestId>(),
-            )
-                .prop_map(|(state, replay, continuation, initial_mast_forest_id)| Self {
-                    state,
-                    replay,
-                    continuation,
-                    initial_mast_forest_id,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for NodeEndData {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (any::<Felt>(), any::<Felt>(), any::<Felt>())
-                .prop_map(|(ended_node_addr, prev_addr, prev_parent_addr)| Self {
-                    ended_node_addr,
-                    prev_addr,
-                    prev_parent_addr,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for ExecutionContextSystemInfo {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (any::<ContextId>(), any::<Word>())
-                .prop_map(|(parent_ctx, parent_fn_hash)| Self { parent_ctx, parent_fn_hash })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for ExecutionContextReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque(any::<ExecutionContextSystemInfo>())
-                .prop_map(|execution_contexts| Self { execution_contexts })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for BlockStackReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (arb_vec_deque(any::<Felt>()), arb_vec_deque(any::<NodeEndData>()))
-                .prop_map(|(node_start_parent_addr, node_end)| Self {
-                    node_start_parent_addr,
-                    node_end,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for MastForestResolutionReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque((any::<MastNodeId>(), any::<MastForestId>()))
-                .prop_map(|mast_forest_resolutions| Self { mast_forest_resolutions })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for MemoryReadsReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                arb_vec_deque((any::<Felt>(), any::<Felt>(), any::<ContextId>(), arb_row_index())),
-                arb_vec_deque((any::<Word>(), any::<Felt>(), any::<ContextId>(), arb_row_index())),
-            )
-                .prop_map(|(elements_read, words_read)| Self { elements_read, words_read })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for MemoryWritesReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                arb_vec_deque((any::<Felt>(), any::<Felt>(), any::<ContextId>(), arb_row_index())),
-                arb_vec_deque((any::<Word>(), any::<Felt>(), any::<ContextId>(), arb_row_index())),
-            )
-                .prop_map(|(elements_written, words_written)| Self {
-                    elements_written,
-                    words_written,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for AdviceReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                arb_vec_deque(any::<Felt>()),
-                arb_vec_deque(any::<Word>()),
-                arb_vec_deque(any::<[Word; 2]>()),
-            )
-                .prop_map(|(stack_pops, stack_word_pops, stack_dword_pops)| Self {
-                    stack_pops,
-                    stack_word_pops,
-                    stack_dword_pops,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for BitwiseOp {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            prop_oneof![Just(Self::U32And), Just(Self::U32Xor)].boxed()
-        }
-    }
-
-    impl Arbitrary for BitwiseReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque((any::<BitwiseOp>(), any::<Felt>(), any::<Felt>()))
-                .prop_map(|u32op_with_operands| Self { u32op_with_operands })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for KernelReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque(any::<Word>())
-                .prop_map(|kernel_proc_accesses| Self { kernel_proc_accesses })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for RangeCheckerReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque(any::<[u16; 4]>())
-                .prop_map(|range_checks_u32_ops| Self { range_checks_u32_ops })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for BlockAddressReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque(any::<Felt>())
-                .prop_map(|block_addresses| Self { block_addresses })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for HasherResponseReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                arb_vec_deque((any::<Felt>(), any::<[Felt; STATE_WIDTH]>())),
-                arb_vec_deque((any::<Felt>(), any::<Word>())),
-                arb_vec_deque((any::<Felt>(), any::<Word>(), any::<Word>())),
-            )
-                .prop_map(
-                    |(
-                        permutation_operations,
-                        build_merkle_root_operations,
-                        mrupdate_operations,
-                    )| {
-                        Self {
-                            permutation_operations,
-                            build_merkle_root_operations,
-                            mrupdate_operations,
-                        }
-                    },
-                )
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for HasherOp {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            prop_oneof![
-                any::<[Felt; STATE_WIDTH]>().prop_map(Self::Permute),
-                (any::<Word>(), any::<Word>(), any::<Felt>(), any::<Word>())
-                    .prop_map(Self::HashControlBlock),
-                (any::<MastForestId>(), any::<MastNodeId>(), any::<Word>())
-                    .prop_map(Self::HashBasicBlock),
-                (any::<Word>(), arb_merkle_path(), any::<Felt>()).prop_map(Self::BuildMerkleRoot),
-                (any::<Word>(), any::<Word>(), arb_merkle_path(), any::<Felt>())
-                    .prop_map(Self::UpdateMerkleRoot),
-            ]
-            .boxed()
-        }
-    }
-
-    impl Arbitrary for HasherRequestReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque(any::<HasherOp>())
-                .prop_map(|hasher_ops| Self { hasher_ops })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for StackOverflowReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                arb_vec_deque((any::<Felt>(), any::<Felt>())),
-                arb_vec_deque((MIN_STACK_DEPTH..=MIN_STACK_DEPTH + 64, any::<Felt>())),
-            )
-                .prop_map(|(overflow_values, restore_context_info)| Self {
-                    overflow_values,
-                    restore_context_info,
-                })
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for ExecutionReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                any::<BlockStackReplay>(),
-                any::<ExecutionContextReplay>(),
-                any::<StackOverflowReplay>(),
-                any::<MemoryReadsReplay>(),
-                any::<AdviceReplay>(),
-                any::<HasherResponseReplay>(),
-                any::<BlockAddressReplay>(),
-                any::<MastForestResolutionReplay>(),
-            )
-                .prop_map(
-                    |(
-                        block_stack,
-                        execution_context,
-                        stack_overflow,
-                        memory_reads,
-                        advice,
-                        hasher,
-                        block_address,
-                        mast_forest_resolution,
-                    )| Self {
-                        block_stack,
-                        execution_context,
-                        stack_overflow,
-                        memory_reads,
-                        advice,
-                        hasher,
-                        block_address,
-                        mast_forest_resolution,
-                    },
-                )
-                .boxed()
-        }
-    }
-
-    impl Arbitrary for AceReplay {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            arb_vec_deque((arb_row_index(), any::<CircuitEvaluation>()))
-                .prop_map(|circuit_evaluations| Self { circuit_evaluations })
-                .boxed()
-        }
-    }
-}
+mod arbitrary;
 
 #[cfg(test)]
-mod serialization_tests {
-    use alloc::vec::Vec;
-
-    use miden_core::serde::{BudgetedReader, SliceReader};
-
-    use super::*;
-    use crate::mast::MastForestId;
-
-    #[test]
-    fn memory_replay_queues_reject_oversized_lengths_before_allocation() {
-        let mut bytes = Vec::new();
-        usize::MAX.write_into(&mut bytes);
-
-        let mut reader = BudgetedReader::new(SliceReader::new(&bytes), bytes.len());
-        let err = read_memory_element_queue(&mut reader).unwrap_err();
-        assert!(err.to_string().contains("exceeds reader allocation bound"));
-
-        let mut reader = BudgetedReader::new(SliceReader::new(&bytes), bytes.len());
-        let err = read_memory_word_queue(&mut reader).unwrap_err();
-        assert!(err.to_string().contains("exceeds reader allocation bound"));
-    }
-
-    #[test]
-    fn stack_state_read_rejects_depth_below_minimum() {
-        let mut bytes = Vec::new();
-        [ZERO; MIN_STACK_DEPTH].write_into(&mut bytes);
-        (MIN_STACK_DEPTH - 1).write_into(&mut bytes);
-        ZERO.write_into(&mut bytes);
-
-        let err = StackState::read_from_bytes(&bytes).unwrap_err();
-        assert!(err.to_string().contains("below minimum"));
-    }
-
-    #[test]
-    fn execution_replay_round_trip_preserves_replay_order() {
-        let mut replay = ExecutionReplay::default();
-        replay.block_stack.record_node_start_parent_addr(Felt::from_u32(7));
-        replay
-            .block_stack
-            .record_node_end(Felt::from_u32(9), Felt::from_u32(8), Felt::from_u32(7));
-        replay.execution_context.record_execution_context(ExecutionContextSystemInfo {
-            parent_ctx: ContextId::from(3),
-            parent_fn_hash: Word::from([ONE, ZERO, ZERO, ZERO]),
-        });
-        replay
-            .stack_overflow
-            .record_pop_overflow(Felt::from_u32(11), Felt::from_u32(10));
-        replay
-            .stack_overflow
-            .record_restore_context_overflow_addr(17, Felt::from_u32(16));
-        replay.memory_reads.record_read_element(
-            Felt::from_u32(21),
-            Felt::from_u32(20),
-            ContextId::from(4),
-            RowIndex::from(5_u32),
-        );
-        replay.memory_reads.record_read_word(
-            Word::from([
-                Felt::from_u32(1),
-                Felt::from_u32(2),
-                Felt::from_u32(3),
-                Felt::from_u32(4),
-            ]),
-            Felt::from_u32(24),
-            ContextId::from(4),
-            RowIndex::from(6_u32),
-        );
-        replay.advice.record_pop_stack(Felt::from_u32(31));
-        replay
-            .advice
-            .record_pop_stack_word(Word::from([Felt::from_u32(5), ZERO, ZERO, ZERO]));
-        replay.advice.record_pop_stack_dword([
-            Word::from([Felt::from_u32(6), ZERO, ZERO, ZERO]),
-            Word::default(),
-        ]);
-        replay.hasher.record_permute(Felt::from_u32(40), [Felt::from_u32(41); 12]);
-        replay.block_address.record_block_address(Felt::from_u32(50));
-        replay
-            .mast_forest_resolution
-            .record_resolution(MastNodeId::from(2), MastForestId::from(1));
-
-        let bytes = replay.to_bytes();
-        let mut restored = ExecutionReplay::read_from_bytes(&bytes).unwrap();
-
-        assert_eq!(
-            restored.block_stack.replay_node_start_parent_addr().unwrap(),
-            Felt::from_u32(7)
-        );
-        let node_end = restored.block_stack.replay_node_end().unwrap();
-        assert_eq!(node_end.ended_node_addr, Felt::from_u32(9));
-        assert_eq!(node_end.prev_addr, Felt::from_u32(8));
-        assert_eq!(node_end.prev_parent_addr, Felt::from_u32(7));
-        let ctx = restored.execution_context.replay_execution_context().unwrap();
-        assert_eq!(ctx.parent_ctx, ContextId::from(3));
-        assert_eq!(ctx.parent_fn_hash, Word::from([ONE, ZERO, ZERO, ZERO]));
-        assert_eq!(
-            restored.stack_overflow.replay_pop_overflow().unwrap(),
-            (Felt::from_u32(11), Felt::from_u32(10))
-        );
-        assert_eq!(
-            restored.stack_overflow.replay_restore_context_overflow_addr().unwrap(),
-            (17, Felt::from_u32(16))
-        );
-        assert_eq!(
-            restored.memory_reads.replay_read_element(Felt::from_u32(20)).unwrap(),
-            Felt::from_u32(21)
-        );
-        assert_eq!(
-            restored.memory_reads.replay_read_word(Felt::from_u32(24)).unwrap(),
-            Word::from([
-                Felt::from_u32(1),
-                Felt::from_u32(2),
-                Felt::from_u32(3),
-                Felt::from_u32(4)
-            ])
-        );
-        assert_eq!(restored.advice.replay_pop_stack().unwrap(), Felt::from_u32(31));
-        assert_eq!(
-            restored.advice.replay_pop_stack_word().unwrap(),
-            Word::from([Felt::from_u32(5), ZERO, ZERO, ZERO])
-        );
-        assert_eq!(
-            restored.advice.replay_pop_stack_dword().unwrap(),
-            [Word::from([Felt::from_u32(6), ZERO, ZERO, ZERO]), Word::default()]
-        );
-        assert_eq!(
-            restored.hasher.replay_permute().unwrap(),
-            (Felt::from_u32(40), [Felt::from_u32(41); 12])
-        );
-        assert_eq!(restored.block_address.replay_block_address().unwrap(), Felt::from_u32(50));
-        assert_eq!(
-            restored.mast_forest_resolution.replay_resolution().unwrap(),
-            (MastNodeId::from(2), MastForestId::from(1))
-        );
-    }
-}
+mod serialization_tests;
